Fixed GLM in the other projects and changed the audio buffer from int to float

+++ LambdaEngine/Include/Audio/Lambda/SoundEffect3DLambda.h
@@ -17,13 +17,13 @@ namespace LambdaEngine
 		virtual bool Init(const SoundEffect3DDesc& desc) override final;
 		virtual void PlayOnceAt(const glm::vec3& position, const glm::vec3& velocity, float volume, float pitch) override final;
 
-		FORCEINLINE const byte* GetWaveform()	{ return m_pWaveForm; }
+		FORCEINLINE const float32* GetWaveform()	{ return m_pWaveForm; }
 		FORCEINLINE uint32 GetSampleCount()		{ return m_Header.SampleCount; }
 		FORCEINLINE uint32 GetSampleRate()		{ return m_Header.SampleRate; }
 		FORCEINLINE uint32 GetChannelCount()	{ return m_Header.ChannelCount; }
 		
 	private:		
-		byte* m_pWaveForm;
+		float32* m_pWaveForm;
 		WaveFile m_Header;
 	};
 }
